requirements.txt
fastapi
openai
python-dotenv
streamlit
matplotlib